+++ src/yalc.ts
@@ -9,6 +9,9 @@ import {
   removePackages,
   getStoreMainDir
 } from '.'
+
+import { showInstallations, cleanInstallations } from './installations'
+
 import { checkManifest } from './check'
 
 const cliCommand = values.myNameIs
@@ -46,6 +49,27 @@ yargs
       })
     }
   })
+  .command({
+    command: 'installations',
+    builder: () => {
+      return yargs
+        .boolean(['dry'])
+    },
+    handler: argv => {      
+      const action = argv._[1]
+      const packages = argv._.slice(2)
+      switch (action) {
+        case 'show':
+          showInstallations({ packages })
+          break
+        case 'clean':
+          cleanInstallations({ packages, dry: argv.dry })
+          break
+        default:
+          console.log('Need installation action: show | clean')
+      }
+    }
+  })  
   .command({
     command: 'push',
     describe: